requirements.txt
aiohttp==3.12.14
aiosqlite==0.21.0
curl_cffi==0.12.0
eth_account==0.13.7
eth_typing==5.2.1
fake_useragent==2.2.0
hexbytes==1.3.1
loguru==0.7.3
python-dotenv==1.1.1
Requests==2.32.4
rich==14.0.0
SQLAlchemy==2.0.41
web3==7.12.1
